Divide before narrowing in Rational's integer conversions
operator int and operator int64_t narrowed the numerator and the
denominator separately and divided those. Normalization is deferred, so
the parts a value is written with are not the ones a caller asked about:
5 can be stored as 15000000000/3000000000, and narrowing each part to int
first turned that into -1. The same fraction over int64_t narrowed the
same way. Both are silent.

    Rational<BigInt>(1, 3000000000) * Rational<BigInt>(15000000000, 1)
        == 5, and numerator()/denominator() agree -- they reduce on
        access -- but static_cast<int> gave -1

    the same product over Rational<int64_t>              int(...) = -1

    Rational<BigInt>(-7, 3000000000) * Rational<BigInt>(3000000000, 2)
        == -7/2                                          int(...) = 0

Dividing in Int and narrowing the quotient is what operator Int already
does, and its comment already states why it is right: den > 0 is an
invariant, so the truncating division num / den matches truncating the
exact value whether or not the fraction is stored reduced. The exact
value fits the narrow type or it does not; how it happens to be written
is not the caller's business.

The floating-point conversions keep dividing after the cast. There the
current form is the accurate one -- dividing first would truncate to an
integer -- and a deferred fraction is capped at about 500 bits, far
inside what long double represents.

Found through the region booleans, whose arrangement vertices are exactly
such deferred fractions. region.intersection<int>(halfplane) reported
vertices nowhere near the crossings: the annulus met with {y >= 5} came
back with (0,5) and (10,5) both collapsed to (0,0).

+++ include/core/rational.hpp
@@ -315,14 +315,23 @@ public:
         return static_cast<long double>(num) / static_cast<long double>(den);
     }
 
-    /// @brief Convert to int
+    /// @brief Convert (truncating toward zero) to `int`.
+    ///
+    /// The division happens in `Int` and only the quotient is narrowed. Doing it
+    /// the other way round — narrowing `num` and `den` and dividing those —
+    /// would be wrong for every fraction stored unreduced: normalization is
+    /// deferred, so a value as ordinary as 5 can be carried as a ratio of two
+    /// numbers far past `int`, and narrowing each of them independently turns
+    /// the quotient into noise. The exact value fits or it does not; the parts
+    /// it happens to be written with are not the caller's business.
     explicit constexpr operator int() const {
-        return static_cast<int>(num) / static_cast<int>(den);
+        return static_cast<int>(num / den);
     }
 
-    /// @brief Convert to int64_t
+    /// @brief Convert (truncating toward zero) to `int64_t`. Divides first, for
+    /// the reason given on @ref operator int() const.
     explicit constexpr operator int64_t() const {
-        return static_cast<int64_t>(num) / static_cast<int64_t>(den);
+        return static_cast<int64_t>(num / den);
     }
 
     /// @brief Convert (truncating toward zero) to the numerator integer type,
